fix: minor changes

+++ src/config/HttpConfig.cpp
@@ -2,19 +2,19 @@
 
 namespace config {
 
-HttpConfig::HttpConfig(size_t keepalive_timeout, size_t client_max_body_size,
+HttpConfig::HttpConfig(unsigned int keepalive_timeout, unsigned int client_max_body_size,
                        const std::map<int, std::string>& error_pages,
                        const std::vector<ServerConfig>& server_configs)
     : keepalive_timeout_(keepalive_timeout), client_max_body_size_(client_max_body_size),
       error_pages_(InitErrorPages(error_pages)), server_configs_(server_configs)
 {}
 
-size_t HttpConfig::keepalive_timeout() const
+unsigned int HttpConfig::keepalive_timeout() const
 {
     return keepalive_timeout_;
 }
 
-size_t HttpConfig::client_max_body_size() const
+unsigned int HttpConfig::client_max_body_size() const
 {
     return client_max_body_size_;
 }
